Initial Commit
I am building an installer for debian and ubuntu along with some customization that I like.

I was going to wait for the installer to be functional before I committed but it's quickly growing very large so I've decided to go ahead and start making commits.

I will probably start tagging with version numbers once the program actually is functional.

The i3 configs will be for a special option I am going to implement to run kde with i3 as the window manager.