Package Change
Biggest change here is the package style formatting. I'm also working on overhauling the way that this whole thing even works. Making dedicated files to split up functions since there is a lot more to do than I originally anticipated.

Next couple of changes will probably be big like this one, but the main file will be easier to read and we'll likely be more modular.

Some other things that I'm going to be doing is converting most if not all paths to pathlib objects. I also may pull the parser into a separate file, Not sure yet.