Very Large Update
We're still not fully operational but we're well on our way. I believe I'm done with the major rewrite.

I'm getting stuck at the sfdisk for right now, but I'm out of time for today so I'm just going to make the commit and then work on it again later.