Small Update
Went through testing and fixed some formatting things. Fixed a couple of things that were typos and stuff causing us to not make it through.
Added an exit() in the main file so that I know where we're making it to. We at least partition and get past debootstrapping which is really nice.
Fixed some filepaths and stuff in the constant file. Fixed a couple of typos in the partition files.
I need to create a chroot function because we need a simple easy way to call commands into the chroot. I also need to make lvcreate more flexible, and potentially sfdisk.
finally cleaned up imports at the top of the files