Partitions as an Object
Changed the partitions into an object. Instead of being a tuple of tuples, it's not a tuple of Custom Partition Objects.

The Class is defined in partition.py. It really makes things easier to read and understand. Also I think this way will be easier to extend functionality in the future.