Functionality Update
Had to make some changes to make the program run with the new Partition Object.

Changed around the part_list tuple so now it's root, boot, esp. This means that when we go to write the fstab file we can just iterate normally and cut out a bunch of logic.

Cleaned up some imports and added the github link to the source list